missile launch chance implemented

+++ Shayetet6/Shayetet6/Balistic.cs
@@ -6,7 +6,7 @@ namespace Shayetet6
 {
     public class Balistic : Missile
     {
-        public Balistic(string missileType, double chance, int distance) : base(chance, distance)
+        public Balistic(string missileType, int chance, int distance) : base(chance, distance)
         {
             MissileType = "Balistic";
         }
